Revert “Optimize `isConstructor`”
This reverts commit 295c215913b0f1782470dae5d99f17fb9a6bd839.

This optimization could be problematic in cases where `length` is a very high number `x`, for which `new Constructor(x)` that takes a very long time before failing. Also, in case an error is thrown from the constructor itself, the script would fall back to `new Array(x)` with this solution, which shouldn’t happen.

Ref. #3.